requirements.txt
requests
Django
django-heroku
gunicorn
numpy
psycopg2
scipy
Pillow
protobuf
opencv-python
tensorflow
keras
setuptools